moved onTradeExportClick into a separate helper file
